requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pydantic==1.10.5
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
tomli==2.0.1
typing_extensions==4.5.0
yarl==1.8.2
